requirements.txt
# Dependências necessárias para GAIA
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.20.0
huggingface-hub>=0.20.0

aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blis==0.7.9
Bottleneck==1.4.2
Brotli==1.0.9
brotlicffi==1.0.9.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.1
confection==0.1.5
contourpy==1.3.1
cycler==0.11.0
cymem==2.0.6
datasets==3.3.2
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
executing==0.8.3
fastjsonschema==2.20.0
filelock==3.17.0
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gensim==4.3.3
gmpy2==2.2.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface_hub==0.31.2
idna==3.7
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langcodes==3.3.0
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.0
mistune==3.1.2
# mkl_fft==1.3.11
# mkl_random==1.2.8
# mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
murmurhash==1.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numexpr==2.10.1
numpy==1.26.4
overrides==7.4.0
packaging==24.2
pandarallel==1.6.4
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pip==25.1
platformdirs==4.3.7
preshed==3.0.9
prometheus_client==0.21.1
prompt-toolkit==3.0.43
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.0
# pt-core-news-sm==3.7.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6_sip==13.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.6.1
QtPy==2.4.1
referencing==0.30.2
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==78.1.1
shellingham==1.5.0
sip==6.10.0
six==1.17.0
sklearn-compat==0.1.3
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.5.1
stack-data==0.2.0
# sympy==1.13.3
terminado==0.17.1
thinc==8.2.2
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.9.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
unicodedata2==15.1.0
Unidecode==1.4.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
wasabi==0.9.1
wcwidth==0.2.13
weasel==0.3.4
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
widgetsnbextension==4.0.13
wordcloud==1.9.4
xxhash==3.5.0
yarl==1.18.0
